Move shared tests to separate module
